+++ packages/client/src/schemas/table/TableFilter.tsx
@@ -12,6 +12,7 @@ import { SchemaField } from '../../components/schema-renderer';
 import { useTable } from './hooks/useTable';
 import { useCompile } from '../../hooks/useCompile';
 import IconPicker from '../../components/icon-picker';
+import {fieldsToFilterColumns} from './utils';
 
 export const TableFilter = observer((props: any) => {
   const { service } = useTable();
