+++ packages/client/src/schemas/table/hooks/useTable.ts
@@ -1,4 +1,4 @@
-import React, { useContext,createContext, useEffect, useState } from 'react';
+import { useContext } from 'react';
 import {TableContext} from '../context';
 
 export const useTable = () => {
