+++ packages/client/src/schemas/table/types.ts
@@ -1,7 +1,7 @@
 import { ArrayField } from '@formily/core';
-import { FormProvider, observer, RecursionField, Schema, useField, useForm } from '@formily/react';
+import { Schema } from '@formily/react';
 import { BaseResult } from '@ahooksjs/use-request/lib/types';
-import { ListOptions, Resource } from '../../resource';
+import { Resource } from '../../resource';
 
 export interface ITableContext {
   props: any;
